Remove utils.py
