+++ llmfoundry/utils/data_prep_utils.py
@@ -1,6 +1,6 @@
 # Copyright 2022 MosaicML LLM Foundry authors
 # SPDX-License-Identifier: Apache-2.0
-
+import logging 
 import json
 import os
 from glob import glob
@@ -139,3 +139,20 @@ class DownloadingIterable:
                     output_filename=output_filename,
                 )
             yield output_filename
+
+
+def configure_logging(logging_level: str, log: logging.Logger):
+    """Configure logging.
+
+    Args:
+        logging_level (str): Logging level.
+    """
+    logging.basicConfig(
+        format=
+        f'%(asctime)s: [%(process)d][%(threadName)s]: %(levelname)s: %(name)s: %(message)s',
+    )
+    logging_level = logging_level.upper()
+    logging.getLogger('llmfoundry').setLevel(logging_level)
+    logging.getLogger(__name__).setLevel(logging_level)
+    log.info(f'Logging level set to {logging_level}')
+
